@@ -1,25 +1,27 @@
 package edu.isi.oba;
 
-
 import static edu.isi.oba.ObaUtils.get_yaml_data;
-import static org.junit.Assert.*;
+import edu.isi.oba.config.CONFIG_FLAG;
+import edu.isi.oba.config.YamlConfig;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.logging.ConsoleHandler;
 import java.util.logging.Level;
 import java.util.logging.LogManager;
 import java.util.logging.Logger;
 
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+
 import org.semanticweb.owlapi.model.OWLClass;
 import org.semanticweb.owlapi.model.OWLOntologyCreationException;
 
-
-import edu.isi.oba.config.YamlConfig;
 import io.swagger.v3.oas.models.media.ArraySchema;
 import io.swagger.v3.oas.models.media.ComposedSchema;
 import io.swagger.v3.oas.models.media.ObjectSchema;
@@ -27,6 +29,15 @@ import io.swagger.v3.oas.models.media.Schema;
 
 public class RestrictionsTest {
 	static Logger logger = null;
+
+	// Convenience variable so we don't need to retype this for each MapperSchema constructor.
+	private Map<CONFIG_FLAG, Boolean> configFlags = new HashMap<>(){{
+		put(CONFIG_FLAG.ALWAYS_GENERATE_ARRAYS, true);
+		put(CONFIG_FLAG.DEFAULT_DESCRIPTIONS, true);
+		put(CONFIG_FLAG.DEFAULT_PROPERTIES, true);
+		put(CONFIG_FLAG.FOLLOW_REFERENCES, true);
+		put(CONFIG_FLAG.REQUIRED_PROPERTIES_FROM_CARDINALITY, false);
+	}};
 	
 	/**
 	 * This method allows you to configure the logger variable that is required to print several 
@@ -56,16 +67,16 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#University");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("hasRector");	       	        
-			if (property instanceof io.swagger.v3.oas.models.media.ArraySchema) {	        	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("hasRector");
+			if (property instanceof io.swagger.v3.oas.models.media.ArraySchema) {
 				Integer maxItems = ((ArraySchema) property).getMaxItems();
-				assertEquals(expectedResult,maxItems);
-			}			
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+				Assertions.assertEquals(expectedResult, maxItems);
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
 		}
 	}
 	
@@ -73,32 +84,32 @@ public class RestrictionsTest {
 	 * This test attempts to get the OAS representation of a ObjectUnionOf restriction.
 	 */
 	@Test
-	public void testObjectUnionOf() throws OWLOntologyCreationException, Exception {		
+	public void testObjectUnionOf() throws OWLOntologyCreationException, Exception {
 		List<String> expectedResult = new ArrayList<String>();
 		expectedResult.add("#/components/schemas/Organization");
 		expectedResult.add("#/components/schemas/Person");
-		try {	
+		try {
 			this.initializeLogger();
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#StudyMaterial");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("author");		        
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("author");
+			if (property instanceof ArraySchema) {
 				Schema items = ((ArraySchema) property).getItems();
 				List<Schema> itemsValue;
 				if (items instanceof ComposedSchema) {
-					itemsValue =((ComposedSchema) items).getAnyOf();
-					for (int i=0; i<itemsValue.size(); i++ ) {
+					itemsValue = ((ComposedSchema) items).getAnyOf();
+					for (int i = 0; i < itemsValue.size(); i++) {
 						String ref = itemsValue.get(i).get$ref();
-						assertEquals(ref,expectedResult.get(i));
-					}	        	
-				}	
-			} 
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+						Assertions.assertEquals(expectedResult.get(i), ref);
+					}
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
 		}
 	}
 	
@@ -107,7 +118,7 @@ public class RestrictionsTest {
 	 */
 	@Test
 	public void testObjectIntersectionOf() throws OWLOntologyCreationException, Exception {
-		List<String> expectedResult = new ArrayList<String>();			
+		List<String> expectedResult = new ArrayList<String>();
 		expectedResult.add("#/components/schemas/Assignment");
 		expectedResult.add("#/components/schemas/Exam");
 		
@@ -116,23 +127,23 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Course");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("hasEvaluationMethod");		        
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("hasEvaluationMethod");
+			if (property instanceof ArraySchema) {
 				Schema items = ((ArraySchema) property).getItems();
 				List<Schema> itemsValue;
 				if (items instanceof ComposedSchema) {
-					itemsValue =((ComposedSchema) items).getAllOf();
-					for (int i=0; i<itemsValue.size(); i++ ) {
+					itemsValue = ((ComposedSchema) items).getAllOf();
+					for (int i = 0; i < itemsValue.size(); i++) {
 						String ref = itemsValue.get(i).get$ref();
-						assertEquals(ref,expectedResult.get(i));
-					}	        	
-				}	
-			} 
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+						Assertions.assertEquals(expectedResult.get(i), ref);
+					}
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
 		}
 	}
 	
@@ -144,20 +155,21 @@ public class RestrictionsTest {
 	public void testSimpleObjectSomeValuesFrom() throws OWLOntologyCreationException, Exception {
 		try {
 			this.initializeLogger();
-			Boolean expectedResult=false;
+			Boolean expectedResult = false;
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#University");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("hasDepartment");		        
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("hasDepartment");
+
+			if (property instanceof ArraySchema) {
 				Boolean nullable = ((ArraySchema) property).getNullable();
-				assertEquals(nullable,expectedResult);					
-			} 
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+				Assertions.assertEquals(expectedResult, nullable);
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
 		}
 	}
 
@@ -167,7 +179,7 @@ public class RestrictionsTest {
 	 */
 	@Test
 	public void testObjectSomeValuesFrom_ComposedByRestriction() throws OWLOntologyCreationException, Exception {
-		List<String> expectedResult = new ArrayList<String>();			
+		List<String> expectedResult = new ArrayList<String>();
 		expectedResult.add("#/components/schemas/BachelorProgram");
 		expectedResult.add("#/components/schemas/MasterProgram");
 		expectedResult.add("#/components/schemas/PhDProgram");
@@ -176,56 +188,105 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Student");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("enrolledIn");		        
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("enrolledIn");
+			if (property instanceof ArraySchema) {
 				Boolean nullable = ((ArraySchema) property).getNullable();
 				Schema items = ((ArraySchema) property).getItems();
 				List<Schema> itemsValue;
 				if (items instanceof ComposedSchema) {
-					itemsValue =((ComposedSchema) items).getAnyOf();
-					for (int i=0; i<itemsValue.size(); i++ ) {			
-						assertEquals(itemsValue.get(i).get$ref(),expectedResult.get(i));
-					}	        	
-				}	
-				assertEquals(nullable,false);					
-			} 			
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
-		}	
-
+					itemsValue = ((ComposedSchema) items).getAnyOf();
+					for (int i = 0; i < itemsValue.size(); i++) {
+						Assertions.assertEquals(expectedResult.get(i), itemsValue.get(i).get$ref());
+					}
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
-	
+
 	/**
-	 * This test attempts to get the OAS representation of the exact cardinality of an ObjectProperty.
+	 * This test attempts to get the OAS representation of the exact cardinality of an ObjectProperty,
+	 * when arrays are set to always be generated for properties.
 	 */
 	@Test
-	public void testObjectExactCardinality() throws OWLOntologyCreationException, Exception {
+	public void testObjectExactCardinalityWithArraysGenerated() throws OWLOntologyCreationException, Exception {
 		try {
 			this.initializeLogger();
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#AmericanStudent");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("hasRecord");		        
-			if (property instanceof ArraySchema) {					
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+
+			this.configFlags.put(CONFIG_FLAG.ALWAYS_GENERATE_ARRAYS, true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("hasRecord");
+			if (property instanceof ArraySchema) {
 				Integer maxItems = ((ArraySchema) property).getMaxItems();
 				Integer minItems = ((ArraySchema) property).getMinItems();
-				if (maxItems!=null && minItems!=null) {
-					if (maxItems == minItems)
-						assertTrue("Exact cardinality configured", true);
-					else
-						assertTrue("Error in exact cardinality restriction", false);
-				} else
-					assertTrue("Null values in exact cardinality restriction.", false);								
-			} 			
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	  
+				if (maxItems != null && minItems != null) {
+					Assertions.assertEquals(1, minItems);
+					Assertions.assertEquals(minItems, maxItems);
+				} else {
+					Assertions.fail("Null values in exact cardinality restriction.");
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
+	}
+	
+	/**
+	 * This test attempts to get the OAS representation of the exact cardinality of an ObjectProperty,
+	 * when properties may or may not be arrays, depending on cardinality.  Plus, list of required properties are set to be generated for schemas.
+	 */
+	@Test
+	public void testObjectExactCardinalityWithRequiredPropertiesAndWithoutArraysGenerated() throws OWLOntologyCreationException, Exception {
+		try {
+			this.initializeLogger();
+			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
+			Mapper mapper = new Mapper(config_data);
+			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#AmericanStudent");
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+
+			this.configFlags.put(CONFIG_FLAG.ALWAYS_GENERATE_ARRAYS, false);
+			this.configFlags.put(CONFIG_FLAG.REQUIRED_PROPERTIES_FROM_CARDINALITY, true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+
+			boolean isRequired = schema.getRequired() != null && schema.getRequired().contains("hasRecord");
+
+			// For exact cardinality, the class schema should have it marked as required when required properties are generated.
+			if (isRequired) {
+				Object property = schema.getProperties().get("hasRecord");
+				Integer maxItems = null;
+				Integer minItems = null;
+				
+				// If the property is an array, it is an object property (i.e. has a "$ref" value).  Otherwise, it is a data property.
+				if (property instanceof ArraySchema) {
+					maxItems = ((ArraySchema) property).getMaxItems();
+					minItems = ((ArraySchema) property).getMinItems();
+				} else if (property instanceof ObjectSchema) {
+					maxItems = ((ObjectSchema) property).getMaxItems();
+					minItems = ((ObjectSchema) property).getMinItems();
+				}
+
+				// Property is known to be required.  And cardinality restrictions should have also been removed, so min/max items should be null.
+				if (minItems == null && maxItems == null) {
+					return;
+				} else {
+					Assertions.fail("Error in exact cardinality restriction.");
+				}
+			} else {
+				Assertions.fail("Error in exact cardinality restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -239,22 +300,23 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#AmericanStudent");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("takesCourse");		        
-			if (property instanceof ArraySchema) {									
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("takesCourse");
+			if (property instanceof ArraySchema) {
 				Integer minItems = ((ArraySchema) property).getMinItems();
-				if (minItems!=null) {
+				if (minItems != null) {
 					Schema items = ((ArraySchema) property).getItems();
-					assertEquals(items.get$ref(),"#/components/schemas/Course");
-					assertEquals(minItems,expectedResult);
-				} else
-					assertTrue("Wrong values in minimum cardinality restriction.", false);								
-			} 			
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	  
+					Assertions.assertEquals("#/components/schemas/Course", items.get$ref());
+					Assertions.assertEquals(expectedResult, minItems);
+				} else {
+					Assertions.fail("Wrong values in minimum cardinality restriction.");
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -268,25 +330,24 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Course");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("hasStudentEnrolled");		        
-			if (property instanceof ArraySchema) {									
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("hasStudentEnrolled");
+			if (property instanceof ArraySchema) {
 				Integer maxItems = ((ArraySchema) property).getMaxItems();
-				if (maxItems!=null) {
+				if (maxItems != null) {
 					Schema items = ((ArraySchema) property).getItems();
-					assertEquals(items.get$ref(),"#/components/schemas/Student");
-					assertEquals(maxItems,expectedResult);
-				} else
-					assertTrue("Wrong values in maximum cardinality restriction.", false);								
-			} 			
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	
-
+					Assertions.assertEquals("#/components/schemas/Student", items.get$ref());
+					Assertions.assertEquals(expectedResult, maxItems);
+				} else {
+					Assertions.fail("Wrong values in maximum cardinality restriction.");
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
-	
 
 	/**
 	 * This test attempts to get the OAS representation of the complementOf of an ObjectProperty.
@@ -299,42 +360,63 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#ProfessorInOtherDepartment");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	
-			if (schema.getNot()!=null)
-				assertEquals(schema.getNot().get$ref(),expectedResult);				
-			else
-				assertTrue("Wrong configuration of ComplementOf restriction", false);
-				
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			if (schema.getNot() != null) {
+				Assertions.assertEquals(expectedResult, schema.getNot().get$ref());
+			} else {
+				Assertions.fail("Wrong configuration of ComplementOf restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
 	 * This test attempts to get the OAS representation of the hasValue of an ObjectProperty.
+	 * 
+	 * This corresponds to a named individual (of a particular class).  Before version 3.1.0, OpenAPI does not support
+	 * the "$ref" at the same level of a default value.  For this reason, they are treated as separate items
+	 * under the "allOf" key (which is an array type).  The "$ref" value is the named individual's class and the "default" value is
+	 * the named individual's name.
 	 */
 	@Test
 	public void testObjectHasValue() throws OWLOntologyCreationException, Exception {
 		try {
 			this.initializeLogger();
-			String expectedResult = "<https://w3id.org/example/resource/Department/ArtificialIntelligenceDepartment>";
+			// The short form of: "<https://w3id.org/example/resource/Department/ArtificialIntelligenceDepartment>"
+			// resolves to the default value of: "ArtificialIntelligenceDepartment".
+			String expectedResult = "ArtificialIntelligenceDepartment";
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#ProfessorInArtificialIntelligence");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("belongsTo");	
-			if (((ObjectSchema) property).getDefault()!=null)
-				assertEquals(((ObjectSchema) property).getDefault(),expectedResult);				
-			else
-				assertTrue("Wrong configuration of ObjectHasValue restriction", false);
-				
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	    
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("belongsTo");
+
+			if (property instanceof ArraySchema) {
+				Schema items = ((ArraySchema) property).getItems();
+				if (items != null && (items.getAllOf() != null && !items.getAllOf().isEmpty())) {
+					for (var item: items.getAllOf()) {
+						if (((ObjectSchema) item).getDefault() != null) {
+							Assertions.assertEquals(expectedResult, ((ObjectSchema) item).getDefault());
+							return;
+						}
+					}
+
+					// If we reach here, then none of the items had a default value, so fail.
+					Assertions.fail("Wrong configuration of ObjectHasValue restriction.");
+				} else {
+					Assertions.fail("Wrong configuration of ObjectHasValue restriction.");
+				}
+			} else {
+				Assertions.fail("Wrong configuration of ObjectHasValue restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -350,29 +432,27 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Professor");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("hasDegree");		        
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("hasDegree");
 
-			if (property instanceof ArraySchema) {	
+			if (property instanceof ArraySchema) {
 				Schema items = ((ArraySchema) property).getItems();
 				List<Object> itemsValue;
 				if (items instanceof ComposedSchema) {
-					itemsValue =((ComposedSchema) items).getEnum();
-					for (int i=0; i<itemsValue.size(); i++ ) {
+					itemsValue = ((ComposedSchema) items).getEnum();
+					for (int i = 0; i < itemsValue.size(); i++) {
 						Object ref = itemsValue.get(i);
-						assertEquals(ref.toString(),expectedResult.get(i));
-					}	        	
-				}	
-			} 
-
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	
+						Assertions.assertEquals(expectedResult.get(i), ref.toString());
+					}
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 
-	
 	/**
 	 * This test attempts to get the OAS representation of a FunctionalDataProperty.
 	 */
@@ -384,16 +464,16 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#AmericanStudent");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("birthDate");	       	        
-			if (property instanceof io.swagger.v3.oas.models.media.ArraySchema) {	        	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("birthDate");
+			if (property instanceof io.swagger.v3.oas.models.media.ArraySchema) {
 				Integer maxItems = ((ArraySchema) property).getMaxItems();
-				assertEquals(expectedResult,maxItems);
-			}			
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+				Assertions.assertEquals(expectedResult, maxItems);
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("error in ontology creation: ", e);
 		}
 	}
 	
@@ -411,23 +491,25 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Course");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("ects");		        
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("ects");
+			if (property instanceof ArraySchema) {
 				Schema items = ((ArraySchema) property).getItems();
 				List<Schema> itemsValue;
 				if (items instanceof ComposedSchema) {
-					itemsValue =((ComposedSchema) items).getAnyOf();
-					for (int i=0; i<itemsValue.size(); i++ ) {
-						String ref = itemsValue.get(i).getType();
-						assertEquals(ref,expectedResult.get(i));
-					}	        	
-				}	
-			} 
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+					itemsValue = ((ComposedSchema) items).getAnyOf();
+					if (itemsValue != null) {
+						for (int i = 0; i < itemsValue.size(); i++) {
+							String ref = itemsValue.get(i).getType();
+							Assertions.assertEquals(expectedResult.get(i), ref);
+						}
+					}
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("error in ontology creation: ", e);
 		}
 	}
 
@@ -445,30 +527,31 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#ProfessorInArtificialIntelligence");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("memberOfOtherDepartments");		        
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("memberOfOtherDepartments");
+			if (property instanceof ArraySchema) {
 				Schema items = ((ArraySchema) property).getItems();
 				List<Schema> itemsValue;
 				if (items instanceof ComposedSchema) {
-					itemsValue =((ComposedSchema) items).getAllOf();
-					for (int i=0; i<itemsValue.size(); i++ ) {
+					itemsValue = ((ComposedSchema) items).getAllOf();
+					for (int i = 0; i < itemsValue.size(); i++) {
 						String ref = itemsValue.get(i).getType();
-						assertEquals(ref,expectedResult.get(i));
-					}	        	
-				}	
-			} 
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+						Assertions.assertEquals(expectedResult.get(i), ref);
+					}
+				}
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("error in ontology creation: ", e);
 		}
 	}
 	
 	/**
 	 * This test attempts to get the OAS representation of the SomeValuesFrom restriction of a DataProperty.
 	 */
-	@Ignore
+	@Disabled
+	@Test
 	public void testDataSomeValuesFrom() throws OWLOntologyCreationException, Exception {
 		try {
 			this.initializeLogger();
@@ -476,17 +559,17 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#StudyProgram");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("studyProgramName");		        
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("studyProgramName");
+			if (property instanceof ArraySchema) {
 				Boolean nullable = ((ArraySchema) property).getNullable();
-				assertEquals(((ArraySchema) property).getItems().getType(), expectedResult);
-				assertEquals(nullable,false);	
-			} 
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+				Assertions.assertEquals(expectedResult, ((ArraySchema) property).getItems().getType());
+				Assertions.assertEquals(false, nullable);
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("error in ontology creation: ", e);
 		}
 	}
 
@@ -498,31 +581,32 @@ public class RestrictionsTest {
 	public void testDataSomeValuesFrom_ComposedByRestriction() throws OWLOntologyCreationException, Exception {
 		try {
 			this.initializeLogger();
-			List<String> expectedResult = new ArrayList<String>();			
+			List<String> expectedResult = new ArrayList<String>();
 			expectedResult.add("integer");
 			expectedResult.add("integer");
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#ProfessorInArtificialIntelligence");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("memberOfOtherDepartments");		        
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("memberOfOtherDepartments");
+			if (property instanceof ArraySchema) {
 				Boolean nullable = ((ArraySchema) property).getNullable();
 				Schema items = ((ArraySchema) property).getItems();
 				List<Schema> itemsValue;
 				if (items instanceof ComposedSchema) {
-					itemsValue =((ComposedSchema) items).getAllOf();
-					for (int i=0; i<itemsValue.size(); i++ ) {			
-						assertEquals(itemsValue.get(i).getType(),expectedResult.get(i));
-					}	        	
-				}	
-				assertEquals(nullable,false);					
-			} 			
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
-		}	    		
+					itemsValue = ((ComposedSchema) items).getAllOf();
+					for (int i = 0; i < itemsValue.size(); i++) {
+						Assertions.assertEquals(expectedResult.get(i), itemsValue.get(i).getType());
+					}
+				}
+
+				Assertions.assertEquals(false, nullable);
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -536,15 +620,15 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#StudyProgram");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("studyProgramName");		        
-			if (property instanceof ArraySchema) {	
-				assertEquals(((ArraySchema) property).getItems().getType(), expectedResult);
-			} 
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("error in ontology creation", false);
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("studyProgramName");
+			if (property instanceof ArraySchema) {
+				Assertions.assertEquals(expectedResult, ((ArraySchema) property).getItems().getType());
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
 		}
 	}
 	
@@ -555,32 +639,31 @@ public class RestrictionsTest {
 	public void testDataOneOf() throws OWLOntologyCreationException, Exception {
 		try {
 			this.initializeLogger();
-			List<String> expectedResult = new ArrayList<String>();			
+			List<String> expectedResult = new ArrayList<String>();
 			expectedResult.add("female");
 			expectedResult.add("male");
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Person");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("gender");	
-			if (property instanceof ArraySchema) {	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("gender");
+			if (property instanceof ArraySchema) {
 				Schema items = ((ArraySchema) property).getItems();
 				List<Object> itemsValue;
 				if (items instanceof ComposedSchema) {
-					itemsValue =((ComposedSchema) items).getEnum();
-					for (int i=0; i<itemsValue.size(); i++ ) {			
-						assertEquals(itemsValue.get(i),expectedResult.get(i));
-					}	        	
-				}						
-			} 			
-			else
-				assertTrue("Wrong configuration of DataOneOf restriction", false);
-				
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	    			
+					itemsValue = ((ComposedSchema) items).getEnum();
+					for (int i = 0; i < itemsValue.size(); i++) {
+						Assertions.assertEquals(expectedResult.get(i), itemsValue.get(i));
+					}
+				}
+			} else {
+				Assertions.fail("Wrong configuration of DataOneOf restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -594,19 +677,21 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#AmericanStudent");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("nationality");	
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("nationality");
 			
-			if (((ArraySchema) property).getItems().getDefault()!=null)
-				assertEquals(((ArraySchema) property).getItems().getDefault(),expectedResult);				
-			else
-				assertTrue("Wrong configuration of DataHasValue restriction", false);
-				
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	    			
+			if (property instanceof ArraySchema && ((ArraySchema) property).getItems().getDefault() != null) {
+				Assertions.assertEquals(expectedResult, ((ArraySchema) property).getItems().getDefault());
+			} else if (((Schema) property).getDefault() != null) {
+				Assertions.assertEquals(expectedResult, ((Schema) property).getDefault());
+			} else {
+				Assertions.fail("Wrong configuration of DataHasValue restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -619,21 +704,36 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#University");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("universityName");		        					
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("universityName");
+
+			if (property instanceof ArraySchema) {
 				Integer maxItems = ((ArraySchema) property).getItems().getMaxItems();
 				Integer minItems = ((ArraySchema) property).getItems().getMinItems();
-				if (maxItems!=null && minItems!=null) {
-					if (maxItems == minItems)
-						assertTrue("Exact cardinality configured", true);
-					else
-						assertTrue("Error in exact cardinality restriction", false);
-			} 			
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	    			
+				if (maxItems != null && minItems != null) {
+					if (maxItems == minItems) {
+						// "Exact cardinality configured" -- does this really need to be output for the test?
+						return;
+					} else {
+						Assertions.fail("Error in exact cardinality restriction.");
+					}
+				}
+			} else if (property instanceof Schema) {
+				Integer maxItems = ((Schema) property).getMaxItems();
+				Integer minItems = ((Schema) property).getMinItems();
+				if (maxItems != null && minItems != null) {
+					Assertions.assertEquals(maxItems, minItems);
+				} else {
+					Assertions.fail("Error in exact cardinality restriction.");
+				}
+			} else {
+				Assertions.fail("Wrong configuration of cardinality restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -647,19 +747,20 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Professor");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("researchField");		        					
-			Integer minItems = ((ArraySchema) property).getItems().getMinItems();
-			if (minItems!=null) 
-				assertEquals(minItems,expectedResult);
-			else
-				assertTrue("Error in minimum cardinality restriction", false);
-						
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	    			
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("researchField");
+			Integer minItems = ((ArraySchema) property).getMinItems();
+
+			if (minItems != null) {
+				Assertions.assertEquals(expectedResult, minItems);
+			} else {
+				Assertions.fail("Error in minimum cardinality restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -673,19 +774,20 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Person");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	  	       
-			Object property= schema.getProperties().get("address");		        					
-			Integer maxItems = ((ArraySchema) property).getItems().getMaxItems();
-			if (maxItems!=null) 
-				assertEquals(maxItems,expectedResult);
-			else
-				assertTrue("Error in maximum cardinality restriction", false);
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("address");
+			Integer maxItems = ((ArraySchema) property).getMaxItems();
 
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	    			
+			if (maxItems != null) {
+				Assertions.assertEquals(expectedResult, maxItems);
+			} else {
+				Assertions.fail("Error in maximum cardinality restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
 	}
 	
 	/**
@@ -699,18 +801,20 @@ public class RestrictionsTest {
 			YamlConfig config_data = get_yaml_data("examples/restrictions/config.yaml");
 			Mapper mapper = new Mapper(config_data);
 			OWLClass cls = mapper.manager.getOWLDataFactory().getOWLClass("https://w3id.org/example#Department");
-			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0));
-			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), true);
-			Schema schema = mapperSchema.getSchema();	
-			Object property= schema.getProperties().get("numberOfProfessors");				
-			if (((ArraySchema) property).getItems().getNot()!=null)
-				assertEquals(((ArraySchema) property).getItems().getNot().getType(),expectedResult);				
-			else
-				assertTrue("Wrong configuration of ComplementOf restriction", false);
-				
-		} catch (OWLOntologyCreationException e) {			
-			assertTrue("Error in ontology creation", false);
-		}	    			
-	}
+			String desc = ObaUtils.getDescription(cls, mapper.ontologies.get(0), true);
+			MapperSchema mapperSchema = new MapperSchema(mapper.ontologies, cls, desc, mapper.schemaNames, mapper.ontologies.get(0), this.configFlags);
+			Schema schema = mapperSchema.getSchema();
+			Object property = schema.getProperties().get("numberOfProfessors");
 
+			if (property instanceof ArraySchema && ((ArraySchema) property).getItems().getNot().getType() != null) {
+				Assertions.assertEquals(expectedResult, ((ArraySchema) property).getItems().getNot().getType());
+			} else if (((Schema) property).getNot().getType() != null) {
+				Assertions.assertEquals(expectedResult, ((Schema) property).getNot().getType());
+			} else {
+				Assertions.fail("Wrong configuration of ComplementOf restriction.");
+			}
+		} catch (OWLOntologyCreationException e) {
+			Assertions.fail("Error in ontology creation: ", e);
+		}
+	}
 }
