@@ -1,70 +1,66 @@
 package edu.isi.oba;
 
+import static edu.isi.oba.Oba.logger;
+
 import io.swagger.v3.oas.models.media.*;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static edu.isi.oba.Oba.logger;
-
 class MapperDataProperty {
-  private final HashMap<String, String> dataTypes;
+  private final Map<String, String> dataTypes = Map.ofEntries(
+    Map.entry("ENTITIES", "string"),
+    Map.entry("ENTITY", "string"),
+    Map.entry("ID", "string"),
+    Map.entry("IDREF", "string"),
+    Map.entry("IDREFS", "string"),
+    Map.entry("NCName", "string"),
+    Map.entry("NMTOKEN", "string"),
+    Map.entry("NMTOKENS", "string"),
+    Map.entry("NOTATION", "string"),
+    Map.entry("Name", "string"),
+    Map.entry("QName", "string"),
+    Map.entry("anySimpleType", "string"),
+    Map.entry("anyType", "string"),
+    Map.entry("anyURI", "string"),
+    Map.entry("base64Binary", "string"),
+    Map.entry("boolean", "boolean"),
+    Map.entry("byte", "integer"),
+    Map.entry("date", "string"),
+    Map.entry("dateTime", "dateTime"),
+    Map.entry("dateTimeStamp", "dateTime"),
+    Map.entry("decimal", "number"),
+    Map.entry("double", "number"),
+    Map.entry("duration", "string"),
+    Map.entry("float", "number"),
+    Map.entry("gDay", "string"),
+    Map.entry("gMonth", "string"),
+    Map.entry("gMonthYear", "string"),
+    Map.entry("gYear", "string"),
+    Map.entry("gYearMonth", "string"),
+    Map.entry("hexBinary", "string"),
+    Map.entry("int", "integer"),
+    Map.entry("integer", "integer"),
+    Map.entry("language", "string"),
+    Map.entry("long", "integer"),
+    Map.entry("negativeInteger", "integer"),
+    Map.entry("nonNegativeInteger", "integer"),
+    Map.entry("nonPositiveInteger", "integer"),
+    Map.entry("normalizedString", "string"),
+    Map.entry("positiveInteger", "integer"),
+    Map.entry("short", "integer"),
+    Map.entry("string", "string"),
+    Map.entry("time", "string"),
+    Map.entry("token", "string"),
+    Map.entry("unsignedByte", "integer"),
+    Map.entry("unsignedInt", "integer"),
+    Map.entry("unsignedLong", "integer"),
+    Map.entry("unsignedShort", "integer"),
+    Map.entry("langString", "string"),
+    Map.entry("Literal", "string")
+  );
 
-  private void setDataTypes() {
-    this.dataTypes.put("ENTITIES", "string");
-    this.dataTypes.put("ENTITY", "string");
-    this.dataTypes.put("ID", "string");
-    this.dataTypes.put("IDREF", "string");
-    this.dataTypes.put("IDREFS", "string");
-    this.dataTypes.put("NCName", "string");
-    this.dataTypes.put("NMTOKEN", "string");
-    this.dataTypes.put("NMTOKENS", "string");
-    this.dataTypes.put("NOTATION", "string");
-    this.dataTypes.put("Name", "string");
-    this.dataTypes.put("QName", "string");
-    this.dataTypes.put("anySimpleType", "string");
-    this.dataTypes.put("anyType", "string");
-    this.dataTypes.put("anyURI", "string");
-    this.dataTypes.put("base64Binary", "string");
-    this.dataTypes.put("boolean", "boolean");
-    this.dataTypes.put("byte", "integer");
-    this.dataTypes.put("date", "string");
-    this.dataTypes.put("dateTime", "dateTime");
-    this.dataTypes.put("dateTimeStamp", "dateTime");
-    this.dataTypes.put("decimal", "number");
-    this.dataTypes.put("double", "number");
-    this.dataTypes.put("duration", "string");
-    this.dataTypes.put("float", "number");
-    this.dataTypes.put("gDay", "string");
-    this.dataTypes.put("gMonth", "string");
-    this.dataTypes.put("gMonthYear", "string");
-    this.dataTypes.put("gYear", "string");
-    this.dataTypes.put("gYearMonth", "string");
-    this.dataTypes.put("hexBinary", "string");
-    this.dataTypes.put("int", "integer");
-    this.dataTypes.put("integer", "integer");
-    this.dataTypes.put("language", "string");
-    this.dataTypes.put("long", "integer");
-    this.dataTypes.put("negativeInteger", "integer");
-    this.dataTypes.put("nonNegativeInteger", "integer");
-    this.dataTypes.put("nonPositiveInteger", "integer");
-    this.dataTypes.put("normalizedString", "string");
-    this.dataTypes.put("positiveInteger", "integer");
-    this.dataTypes.put("short", "integer");
-    this.dataTypes.put("string", "string");
-    this.dataTypes.put("time", "string");
-    this.dataTypes.put("token", "string");
-    this.dataTypes.put("unsignedByte", "integer");
-    this.dataTypes.put("unsignedInt", "integer");
-    this.dataTypes.put("unsignedLong", "integer");
-    this.dataTypes.put("unsignedShort", "integer");
-    this.dataTypes.put("langString", "string");
-    this.dataTypes.put("Literal", "string");
-
-  }
-
-  private String getDataType(String key){
+  private String getDataType(String key) {
     return this.dataTypes.get(key);
   }
 
@@ -83,10 +79,7 @@ class MapperDataProperty {
   private Map<String,String> restrictions;
   private List<String> valuesFromDataRestrictions_ranges;
 
-
-  public MapperDataProperty(String name, String description, Boolean isFunctional,Map<String,String> restrictions,List<String> valuesFromDataRestrictions_ranges, List<String> type, Boolean array, Boolean nullable) {
-    this.dataTypes = new HashMap<>();
-    this.setDataTypes();
+  public MapperDataProperty(String name, String description, Boolean isFunctional, Map<String, String> restrictions, List<String> valuesFromDataRestrictions_ranges, List<String> type, Boolean array, Boolean nullable) {
     this.name = name;
     this.description = description;
     this.type = type;
@@ -97,33 +90,33 @@ class MapperDataProperty {
     this.valuesFromDataRestrictions_ranges=valuesFromDataRestrictions_ranges;
   }
 
-  public Schema getSchemaByDataProperty(){
+  public Schema getSchemaByDataProperty() {
 	  
-    if (this.type.size() == 0) {
-      return (array) ? arraySchema(new StringSchema(), nullable) : new StringSchema().nullable(nullable).description(description);
-    }
-    else if (this.type.size() > 1){
-    	return (array) ? composedSchema(this.type, nullable) : new Schema().nullable(nullable).description(description);
+    if (this.type.isEmpty()) {
+      return (this.array) ? this.arraySchema(new StringSchema()) : this.nonArraySchema(new StringSchema());
+    } else if (this.type.size() > 1) {
+    	return (this.array) ? this.composedSchema() : this.nonArraySchema(new Schema());
     }
 
     String schemaType = getDataType(this.type.get(0));
-    if (schemaType == null){
+    if (schemaType == null) {
       logger.severe("property " + this.name + " type " + this.type);
     }
+    
     switch (schemaType) {
       case STRING_TYPE:
-        return (array) ? arraySchema(new StringSchema(), nullable) : new StringSchema().nullable(nullable).description(description);
+        return (this.array) ? this.arraySchema(new StringSchema()) : this.nonArraySchema(new StringSchema());
       case NUMBER_TYPE:
-        return (array) ? arraySchema(new NumberSchema(), nullable) : new NumberSchema().nullable(nullable).description(description);
+        return (this.array) ? this.arraySchema(new NumberSchema()) : this.nonArraySchema(new NumberSchema());
       case INTEGER_TYPE:
-        return (array) ? arraySchema(new IntegerSchema(), nullable) : new IntegerSchema().nullable(nullable).description(description);
+        return (this.array) ? this.arraySchema(new IntegerSchema()) : this.nonArraySchema(new IntegerSchema());
       case BOOLEAN_TYPE:
-        return (array) ? arraySchema(new BooleanSchema(), nullable) : new BooleanSchema().nullable(nullable).description(description);
+        return (this.array) ? this.arraySchema(new BooleanSchema()) : this.nonArraySchema(new BooleanSchema());
       case DATETIME_TYPE:
-          return (array) ? arraySchema(new DateTimeSchema() , nullable) : new DateTimeSchema().nullable(nullable).description(description);
+          return (this.array) ? this.arraySchema(new DateTimeSchema()) : this.nonArraySchema(new DateTimeSchema());
       default:
         logger.warning("datatype mapping failed " + this.type.get(0));
-        return (array) ? arraySchema(new Schema(), nullable) : new Schema().nullable(nullable).description(description);
+        return (this.array) ? this.arraySchema(new Schema()) : this.nonArraySchema(new Schema());
     }
   }
 
@@ -132,42 +125,54 @@ class MapperDataProperty {
    *
    * @return An ArraySchema: including a composedSchema with a list of items for anyOf, allOf restrictions.
    */ 
-  private ArraySchema composedSchema(List<String> base, boolean nullable){
+  private ArraySchema composedSchema() {
 	  ArraySchema array = new ArraySchema();
-	  Schema schema ;
-	  ComposedSchema composedSchema = new ComposedSchema() ;
+	  Schema schema;
+	  ComposedSchema composedSchema = new ComposedSchema();
 	  array.setDescription(description);
-	  array.setNullable(nullable);  
+	  array.setNullable(this.nullable);
+
 	  // Operations for managing boolean combinations 
-	  for (String restriction:  restrictions.keySet()) { 
-		  String value = restrictions.get(restriction); 	  
-		  for (String item:base) {		  
-			  switch (getDataType(item)) {
+	  for (String restriction: this.restrictions.keySet()) { 
+		  String value = this.restrictions.get(restriction); 	  
+		  for (String item: this.type) {		  
+			  switch (this.getDataType(item)) {
 			  case STRING_TYPE:
 				  schema = new StringSchema();
-				  if (item.equals("anyURI"))
-					  schema.format("uri");
-				  else if (item.equals("byte"))
-					  schema.format("byte");
-				  break ;
+
+				  if (item.equals("anyURI")) {
+            schema.format("uri");
+          } else if (item.equals("byte")) {
+            schema.format("byte");
+          }
+
+				  break;
 			  case NUMBER_TYPE:
 				  schema = new NumberSchema();
-				  if (item.equals("float"))
-					  schema.format("float");
-				  else if (item.equals("double"))
-					  schema.format("double");
-				  break ;
+
+				  if (item.equals("float")) {
+            schema.format("double");
+          } else if (item.equals("double")) {
+            schema.format("double");
+          } else {
+            schema.format("number");
+          }
+          
+				  break;
 			  case INTEGER_TYPE:
 				  schema = new IntegerSchema();
-				  if (item.equals("long"))
-					  schema.format("int64");
-				  break ;
+
+				  if (item.equals("long")) {
+            schema.format("int64");
+          }
+					  
+				  break;
 			  case BOOLEAN_TYPE:
 				  schema = new BooleanSchema();
-				  break ;
+				  break;
 			  case DATETIME_TYPE:
 				  schema = new DateTimeSchema();
-				  break ;	       
+				  break;	       
 			  default:
 				  logger.warning("datatype mapping failed " + this.type.get(0));
 				  schema = new Schema();	  	
@@ -175,13 +180,17 @@ class MapperDataProperty {
 			  
 			  switch (restriction) {
 			  case "unionOf":
-				  if (value=="someValuesFrom")
-		        		nullable=false;	
+				  if (value == "someValuesFrom") {
+            this.nullable = false;
+          }
+
 				  composedSchema.addAnyOfItem(schema); 
 				  break;
 			  case "intersectionOf":
-				  if (value=="someValuesFrom")
-		        		nullable=false;	
+				  if (value == "someValuesFrom") {
+            this.nullable = false;	
+          }
+
 				  composedSchema.addAllOfItem(schema);
 				  break;
 			  default:
@@ -189,66 +198,102 @@ class MapperDataProperty {
 			  }
 		  }
 	  }
-	  array.setItems(composedSchema);  	 
 
-	  if (isFunctional)
-		  array.setMaxItems(1);
-	  array.setNullable(nullable);	     
-	  return array ;
+	  array.setItems(composedSchema);
+    array.setNullable(this.nullable);
+
+	  if (isFunctional) {
+      array.setMaxItems(1);
+    }
+	     
+	  return array;
   }
 
-  private ArraySchema arraySchema(Schema base, boolean nullable) {
+  private Schema nonArraySchema(Schema base) {
+    base.setDescription(this.description);
+    base.setNullable(this.nullable);
+
+    return this.getSchemaRestrictions(base);
+  }
+
+  private ArraySchema arraySchema(Schema base) {
 	  ArraySchema array = new ArraySchema();
-	  array.setDescription(description);
+	  array.setDescription(this.description);
 
-	  if (isFunctional)
-		  array.setMaxItems(1);
+	  if (this.isFunctional) {
+      array.setMaxItems(1);
+    }
 
-	  for (String restriction:  restrictions.keySet()) { 
-		  String value = restrictions.get(restriction);
-		  switch (restriction) {
-		  case "dataHasValue":      		
-			  base.setDefault(value);
-			  break;
-		  case "maxCardinality":
-			  base.setMaxItems(Integer.parseInt(value));
-			  break ;
-		  case "minCardinality":	
-			  base.setMinItems(Integer.parseInt(value));
-			  break ;
-		  case "exactCardinality":
-			  base.setMaxItems(Integer.parseInt(value));
-			  base.setMinItems(Integer.parseInt(value));
-			  break ;
-		  case "someValuesFrom": 
-			  nullable=false;
-			  break ;
-		  case "allValuesFrom":      	 
-			  //nothing to do
-			  break ; 		   
-		  case "oneOf":   
-			  if (value=="someValuesFrom")
-				  nullable=false;	
-			  for (String rangeValue:valuesFromDataRestrictions_ranges)
-				  base.addEnumItemObject(rangeValue);
-			  break;			 
-		  case "complementOf":  
-			  Schema schema = new Schema();
-			  Schema complementOf = new Schema();
-			  complementOf.setType(getDataType(type.get(0)));
-			  complementOf.setFormat(type.get(0));
-			  schema.setNot(complementOf);
-			  array.setNullable(nullable);
-			  array.setItems(schema);
-			  return array;
-		  default:
-		  } 
-	  } 
+    if (this.restrictions.containsKey("complementOf")) {
+      Schema schema = new Schema();
+      Schema complementOf = new Schema();
+      complementOf.setType(this.getDataType(this.type.get(0)));
+      complementOf.setFormat(this.type.get(0));
+      schema.setNot(complementOf);
+      array.setNullable(this.nullable);
+      array.setItems(schema);
+      return array;
+    }
 
-	  array.setNullable(nullable);
+    base = this.getSchemaRestrictions(base);
+
+    // Can this be done in a better way?
+    Integer minItemsInteger = base.getMinItems();
+    if (minItemsInteger != null) {
+      base.setMinItems(null);
+      array.setMinItems(minItemsInteger);
+    }
+
+    // Can this be done in a better way?
+    Integer maxItemsInteger = base.getMaxItems();
+    if (maxItemsInteger != null) {
+      base.setMaxItems(null);
+      array.setMaxItems(maxItemsInteger);
+    }
+
+	  array.setNullable(this.nullable);
 	  array.setItems(base);
 
 	  return array;
   }
 
+  private Schema getSchemaRestrictions(Schema base) {
+    for (String restriction: this.restrictions.keySet()) { 
+      String value = restrictions.get(restriction);
+      switch (restriction) {
+        case "dataHasValue":
+          base.setDefault(value);
+          break;
+        case "maxCardinality":
+          base.setMaxItems(Integer.parseInt(value));
+          break;
+        case "minCardinality":
+          base.setMinItems(Integer.parseInt(value));
+          break;
+        case "exactCardinality":
+          base.setMaxItems(Integer.parseInt(value));
+          base.setMinItems(Integer.parseInt(value));
+          break;
+        case "someValuesFrom":
+          this.nullable = false;
+          break;
+        case "allValuesFrom":
+          //nothing to do
+          break;
+        case "oneOf":
+          if (value == "someValuesFrom") {
+            this.nullable = false;
+          }
+
+          for (String rangeValue: this.valuesFromDataRestrictions_ranges) {
+            base.addEnumItemObject(rangeValue);
+          }
+
+          break;
+        default:
+      }
+    }
+
+    return base;
+  }
 }
