@@ -1,11 +1,13 @@
 package edu.isi.oba;
 
-import io.swagger.v3.oas.models.media.*;
-
 import static edu.isi.oba.Oba.logger;
 
+import io.swagger.v3.oas.models.media.*;
+
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 class MapperObjectProperty {
   final String name;
@@ -23,7 +25,7 @@ class MapperObjectProperty {
     this.ref = ref;
     this.array = true;
     this.nullable = true;
-    this.isFunctional=isFunctional;
+    this.isFunctional = isFunctional;
     this.restrictions = restrictions;
   }
 
@@ -33,148 +35,196 @@ class MapperObjectProperty {
     this.ref = ref;
     this.array = array;
     this.nullable = nullable;
-    this.isFunctional=isFunctional;
+    this.isFunctional = isFunctional;
     this.restrictions = restrictions;
   }
 
-  public Schema getSchemaByObjectProperty(){
-    if (this.ref.size() == 0){
-      return getComposedSchemaObject(this.ref, array, nullable);
+  public Schema getSchemaByObjectProperty() {
+    if (this.ref.isEmpty()){
+      return getComposedSchemaObject(this.ref, this.array, this.nullable);
     }
 
     if (this.ref.size() > 1){
-      return getComposedSchemaObject(this.ref, array, nullable);
-    }
-    else {
-      return getObjectPropertiesByRef(this.ref.get(0), array, nullable);
+      return getComposedSchemaObject(this.ref, this.array, this.nullable);
+    } else {
+      return getObjectPropertiesByRef(this.ref.get(0), this.array, this.nullable);
     }
   }
 
-  private Schema getObjectPropertiesByRef(String ref, boolean array, boolean nullable){
-    Schema object = new ObjectSchema();
-    object.setDescription(description);
-
-    if (array) {
+  private Schema getObjectPropertiesByRef(String ref, boolean isArray, boolean isNullable) {
+    if (this.restrictions.isEmpty()) {
+      Schema object = new ObjectSchema();
       object.set$ref(ref);
-      ArraySchema objects = new ArraySchema();
-      objects.setDescription(description);
-      if (isFunctional)
+
+      if (isArray) {
+        ArraySchema objects = new ArraySchema();
+        objects.setDescription(this.description);
+        objects.setNullable(isNullable);
+        objects.setItems(object);
+
+        if (this.isFunctional) {
           objects.setMaxItems(1);
-      
-      for (String restriction:  restrictions.keySet()) { 
-    	  String value = restrictions.get(restriction);
-      switch (restriction) {
-      case "maxCardinality":
-    	objects.setMaxItems(Integer.parseInt(value));
-      	break ;
-      case "minCardinality":	
-    	  objects.setMinItems(Integer.parseInt(value));
-      	break ;
-      case "exactCardinality":
-    	  objects.setMaxItems(Integer.parseInt(value));
-    	  objects.setMinItems(Integer.parseInt(value));
-      	break ;
-      case "someValuesFrom": 
-    	  nullable=false;
-      	break ;
-      case "allValuesFrom":      	  
-    	  //nothing to do in the Schema
-      	break ;   
-      default:
-    	  break ; 
-    	} 
-      }     
-      objects.setNullable(nullable);
-      objects.setItems(object);      
-      return objects;
-    }
-    else {
-    	for (String restriction:  restrictions.keySet()) { 
-    		String value = restrictions.get(restriction);
-    		if (restriction=="objectHasValue") {      		
-    			object.setDefault(value);
-    			object.type("string");
-    			object.format("uri");
-    			return object;
-    		}
-    	}
-    	return object;
+        }
+
+        return objects;
+      } else {
+        object.setDescription(this.description);
+        object.setNullable(isNullable);
+        object.setType("object");
+
+        if (this.isFunctional) {
+          object.setMaxItems(1);
+        }
+
+        return object;
+      }
     }
 
-    
+    ArraySchema objects = new ArraySchema();
+    objects.setType("array");
+
+    // For OpenAPI v3.0, "$ref" cannot be used as a sibling with "default".
+    // see: https://swagger.io/docs/specification/using-ref/
+    // see: https://stackoverflow.com/a/77189463
+    // A workaround is to include both within "items" and as separate entries under "allOf".
+    // TODO: version check here (and elsewhere?) to differentiate the schema structure.  For v3.1+, it can support "$ref" and "default" as siblings.
+    ArraySchema allOfItems = new ArraySchema();
+    allOfItems.setType(null);
+
+    for (String restriction:  this.restrictions.keySet()) {
+      String value = this.restrictions.get(restriction);
+
+      switch (restriction) {
+        case "maxCardinality":
+          objects.setMaxItems(Integer.parseInt(value));
+          break;
+        case "minCardinality":
+          objects.setMinItems(Integer.parseInt(value));
+          break;
+        case "exactCardinality":
+          objects.setMaxItems(Integer.parseInt(value));
+          objects.setMinItems(Integer.parseInt(value));
+          break;
+        case "someValuesFrom":
+          isNullable = false;
+          break;
+        case "allValuesFrom":
+          //nothing to do in the Schema
+          break;
+        case "objectHasReference": {
+          Schema object = new ObjectSchema();
+          object.set$ref(value);
+          object.setType(null);
+          allOfItems.addAllOfItem(object);
+          break;
+        }
+        case "objectHasValue": {
+          Schema object = new ObjectSchema();
+          object.setDefault(value);
+          object.setType("string");
+          allOfItems.addAllOfItem(object);
+          break;
+        }
+        default:
+          break;
+      }
+    }
+
+    if (allOfItems.getAllOf() == null || allOfItems.getAllOf().isEmpty()) {
+      Schema object = new ObjectSchema();
+      object.set$ref(ref);
+      objects.items(object);
+    } else {
+      objects.items(allOfItems);
+    }
+
+    if (this.isFunctional) {
+      objects.setMaxItems(1);
+    }
+
+    objects.setNullable(isNullable);
+
+    return objects;
   }
-  private Schema getComposedSchemaObject(List<String> refs, boolean array, boolean nullable){
+
+  private Schema getComposedSchemaObject(List<String> refs, boolean array, boolean nullable) {
     Schema object = new ObjectSchema();
-    ComposedSchema composedSchema = new ComposedSchema() ;
+    ComposedSchema composedSchema = new ComposedSchema();
     
     object.setType("object");
-    object.setDescription(description);
+    object.setDescription(this.description);
 
-    if (array) {
+    if (array && !this.restrictions.isEmpty()) {
     	ArraySchema objects = new ArraySchema();
-    	objects.setDescription(description);
+    	objects.setDescription(this.description);
     	
-    	if (isFunctional)
+    	if (this.isFunctional) {
     		objects.setMaxItems(1);
+      }
 
-    	for (String restriction:  restrictions.keySet()) { 
-    		String value = restrictions.get(restriction); 
+    	for (String restriction:  this.restrictions.keySet()) { 
+        String value = this.restrictions.get(restriction);
 
-    		for (String item:refs) {
+        // In some cases, the duplicate reference may have been added to the list.  We only need to create one $ref item to it.
+        LinkedHashSet<String> uniqueRefs = refs.stream().collect(Collectors.toCollection(LinkedHashSet::new));
+
+    		for (String item: uniqueRefs) {
     			Schema objectRange = new ObjectSchema();
     			objectRange.setType("object");
     			objectRange.set$ref(item);
+
     			switch (restriction) {
     			case "unionOf":
-    				if (value=="someValuesFrom")
-    					nullable=false;	
+    				if ("someValuesFrom".equals(value)) {
+    					nullable=false;
+            }
+
     				composedSchema.addAnyOfItem(objectRange);
     				objects.setItems(composedSchema);
-    				break ;
+    				break;
     			case "intersectionOf":
-    				if (value=="someValuesFrom")
-    					nullable=false;	
+    				if ("someValuesFrom".equals(value)) {
+    					nullable = false;
+            }
+
     				composedSchema.addAllOfItem(objectRange);
     				objects.setItems(composedSchema);
-    				break ;
-    			case "someValuesFrom": 
-    				nullable=false;
     				break;
-    			case "allValuesFrom":      	  
+    			case "someValuesFrom":
+    				nullable=false;
+            composedSchema.addAnyOfItem(objectRange);
+            objects.setItems(composedSchema);
+    				break;
+    			case "allValuesFrom":
     				//nothing to do in the Schema
-    				break ;  
-    	        case "oneOf":  
-    	        	if (value=="someValuesFrom")
-    	        		nullable=false;	
-    	        	composedSchema.addEnumItemObject(item);
-    	            composedSchema.setType("string");
-    	            composedSchema.setFormat("uri");
-    	        	objects.setItems(composedSchema);
-    	        	break;
+    				break;
+    	    case "oneOf":
+            if (value=="someValuesFrom") {
+              nullable=false;
+            }
+
+            composedSchema.addEnumItemObject(item);
+            composedSchema.setType("string");
+            composedSchema.setFormat("uri");
+            objects.setItems(composedSchema);
+            break;
     			default:
     				//if the property range is complex it will be omitted   
-    				logger.warning("omitted complex restriction");	 
-    				objects.setItems(object); 		          
-    			}             
+    				logger.warning("omitted complex restriction");
+    				objects.setItems(object);
+    			}
     		}
-    	}     
-    	if (refs.size() == 0)
-    		objects.setItems(object);
-    		objects.setNullable(nullable);
+    	}
+
+    	if (refs.isEmpty()) {
+        objects.setItems(object);
+      }
+    	objects.setNullable(nullable);
+
     	return objects;
-    }
-    else {
+    } else {
+      object.setNullable(nullable);
       return object;
     }
   }
-
-  private ArraySchema arraySchema(Schema base, boolean nullable) {
-    ArraySchema array = new ArraySchema();
-    array.setNullable(nullable);
-    array.setItems(base);
-    if (isFunctional)
-        array.setMaxItems(1);
-    return array;
-  }
-
 }
