@@ -3,65 +3,79 @@ package edu.isi.oba.config;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.PathItem;
 
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-
 public class YamlConfig {
+  private final Map<CONFIG_FLAG, Boolean> configFlags = new HashMap<>(){{
+    put(CONFIG_FLAG.ALWAYS_GENERATE_ARRAYS, true);
+    put(CONFIG_FLAG.DEFAULT_DESCRIPTIONS, true);
+    put(CONFIG_FLAG.DEFAULT_PROPERTIES, true);
+    put(CONFIG_FLAG.FOLLOW_REFERENCES, true);
+    put(CONFIG_FLAG.PATH_DELETE, false);
+    put(CONFIG_FLAG.PATH_GET, true);
+    put(CONFIG_FLAG.PATH_PATCH, false);
+    put(CONFIG_FLAG.PATH_POST, false);
+    put(CONFIG_FLAG.PATH_PUT, false);
+    put(CONFIG_FLAG.REQUIRED_PROPERTIES_FROM_CARDINALITY, false);
+  }};
+
   String DEFAULT_OUTPUT_DIRECTORY = "outputs";
   String DEFAULT_PROJECT_NAME = "default_project";
   public OpenAPI openapi;
   public String output_dir = DEFAULT_OUTPUT_DIRECTORY;
   public String name = DEFAULT_PROJECT_NAME;
   public List<String> paths;
-  public Boolean enable_get_paths = true;
-  public Boolean enable_post_paths = false;
-  public Boolean enable_put_paths = false;
-  public Boolean enable_delete_paths = false;
   public List<String> ontologies;
   private EndpointConfig endpoint;
-
   public AuthConfig auth;
   public FirebaseConfig firebase;
   public Map<String, List<RelationConfig>> relations;
   private LinkedHashMap<String, PathItem> custom_paths = null;
   public List<String> classes;
-  public Boolean follow_references = false;
+  public String custom_queries_directory;
 
   public Boolean getEnable_get_paths() {
-    return enable_get_paths;
+    return this.configFlags.get(CONFIG_FLAG.PATH_GET);
   }
 
   public void setEnable_get_paths(Boolean enable_get_paths) {
-    this.enable_get_paths = enable_get_paths;
+    this.configFlags.put(CONFIG_FLAG.PATH_GET, enable_get_paths);
+  }
+
+  public Boolean getEnable_patch_paths() {
+    return this.configFlags.get(CONFIG_FLAG.PATH_PATCH);
+  }
+
+  public void setEnable_patch_paths(Boolean enable_patch_paths) {
+    this.configFlags.put(CONFIG_FLAG.PATH_PATCH, enable_patch_paths);
   }
 
   public Boolean getEnable_post_paths() {
-    return enable_post_paths;
+    return this.configFlags.get(CONFIG_FLAG.PATH_POST);
   }
 
   public void setEnable_post_paths(Boolean enable_post_paths) {
-    this.enable_post_paths = enable_post_paths;
+    this.configFlags.put(CONFIG_FLAG.PATH_POST, enable_post_paths);
   }
 
   public Boolean getEnable_put_paths() {
-    return enable_put_paths;
+    return this.configFlags.get(CONFIG_FLAG.PATH_PUT);
   }
 
   public void setEnable_put_paths(Boolean enable_put_paths) {
-    this.enable_put_paths = enable_put_paths;
+    this.configFlags.put(CONFIG_FLAG.PATH_PUT, enable_put_paths);
   }
 
   public Boolean getEnable_delete_paths() {
-    return enable_delete_paths;
+    return this.configFlags.get(CONFIG_FLAG.PATH_DELETE);
   }
 
   public void setEnable_delete_paths(Boolean enable_delete_paths) {
-    this.enable_delete_paths = enable_delete_paths;
+    this.configFlags.put(CONFIG_FLAG.PATH_DELETE, enable_delete_paths);
   }
-
-  public String custom_queries_directory;
 
   public String getCustom_queries_directory() {
     return custom_queries_directory;
@@ -127,7 +141,6 @@ public class YamlConfig {
     this.relations = relations;
   }
 
-
   public LinkedHashMap<String, PathItem> getCustom_paths() {
     return custom_paths;
   }
@@ -144,20 +157,52 @@ public class YamlConfig {
     this.openapi = openapi;
   }
 
-    public List<String> getClasses() {
-      return this.classes;
-    }
+  public List<String> getClasses() {
+    return this.classes;
+  }
 
   public void setClasses(List<String> classes) {
     this.classes = classes;
   }
 
+  public Boolean getAlways_generate_arrays() {
+    return this.configFlags.get(CONFIG_FLAG.ALWAYS_GENERATE_ARRAYS);
+  }
+
+  public void setAlways_generate_arrays(Boolean always_generate_arrays) {
+    this.configFlags.put(CONFIG_FLAG.ALWAYS_GENERATE_ARRAYS, always_generate_arrays);
+  }
+
   public Boolean getFollow_references() {
-    return follow_references;
+    return this.configFlags.get(CONFIG_FLAG.FOLLOW_REFERENCES);
   }
 
   public void setFollow_references(Boolean follow_references) {
-    this.follow_references = follow_references;
+    this.configFlags.put(CONFIG_FLAG.FOLLOW_REFERENCES, follow_references);
+  }
+
+  public Boolean getDefault_descriptions() {
+    return this.configFlags.get(CONFIG_FLAG.DEFAULT_DESCRIPTIONS);
+  }
+
+  public void setDefault_descriptions(Boolean default_descriptions) {
+    this.configFlags.put(CONFIG_FLAG.DEFAULT_DESCRIPTIONS, default_descriptions);
+  }
+
+  public Boolean getDefault_properties() {
+    return this.configFlags.get(CONFIG_FLAG.DEFAULT_PROPERTIES);
+  }
+
+  public void setDefault_properties(Boolean default_properties) {
+    this.configFlags.put(CONFIG_FLAG.DEFAULT_PROPERTIES, default_properties);
+  }
+
+  public Boolean getRequired_properties_from_cardinality() {
+    return this.configFlags.get(CONFIG_FLAG.REQUIRED_PROPERTIES_FROM_CARDINALITY);
+  }
+
+  public void setRequired_properties_from_cardinality(Boolean required_properties_from_cardinality) {
+    this.configFlags.put(CONFIG_FLAG.REQUIRED_PROPERTIES_FROM_CARDINALITY, required_properties_from_cardinality);
   }
 
   public AuthConfig getAuth() {
@@ -167,5 +212,23 @@ public class YamlConfig {
   public void setAuth(AuthConfig auth) {
     this.auth = auth;
   }
-}
 
+  /**
+   * Get the value of a particular configuration flag.
+   * 
+   * @param {flag} the configuration flag name
+   * @return The flag's value (true/false/null).
+   */
+  public Boolean getConfigFlagValue(CONFIG_FLAG flag) {
+    return this.configFlags.get(flag);
+  }
+
+  /**
+   * Get map of all config flags and their values.
+   * 
+   * @return Map of CONFIG_FLAGs and their Boolean value (true/false/null).
+   */
+  public Map<CONFIG_FLAG, Boolean> getConfigFlags() {
+    return this.configFlags;
+  }
+}
